[~] Fixed some headers [~] Moved the old BasicCursor that is only necessary for SPatialUnderstanding example to the SpatialUnderstanding example [~] Fixed a few scenes that had broken references
