+++ Assets/HoloToolkit-Examples/SpatialUnderstanding/SpatialUnderstanding-FeatureOverview/Scripts/SpatialUnderstandingBasicCursor.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 489db1d438a864748936fb2f05c2c6f6
+timeCreated: 1455735877
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
