+++ Assets/HoloToolkit/Input/Scripts/Cursor/ObjectCursor.cs
@@ -1,7 +1,5 @@
-﻿//
-// Copyright (C) Microsoft. All rights reserved.
-// TODO This needs to be validated for HoloToolkit integration
-//
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System;
 using UnityEngine;
